Migrate the form models to their component JC-674
